Complete the master logic
The master will try to acquire the locks of other hosts
  - if they are alive, the master will fail
  - if they fail, the master will succeed after the self-fence time

Once the master has acquired the lock of another host it is guaranteed
that the host will reboot with no filesystems and no VMs. It is therefore
safe for the master to restart all VMs.

The master unlocks the other host's lock, so it can reboot and re-acquire
the lock. The master polls for the lock to be taken. As soon as the
lock is taken the master will try again to acquire it.

Signed-off-by: David Scott <dave.scott@citrix.com>